+++ include/fiwix/ramdisk.h
@@ -27,7 +27,7 @@ int ramdisk_open(struct inode *, struct fd *);
 int ramdisk_close(struct inode *, struct fd *);
 int ramdisk_read(__dev_t, __blk_t, char *, int);
 int ramdisk_write(__dev_t, __blk_t, char *, int);
-int ramdisk_ioctl(struct inode *, int, unsigned long int);
+int ramdisk_ioctl(struct inode *, int, unsigned int);
 __loff_t ramdisk_llseek(struct inode *, __loff_t);
 
 void ramdisk_init(void);
